+++ libraries/P_Kinematics/kinematics_CMP.h
@@ -12,10 +12,10 @@
 
 unsigned long kinematics_timer;
 
-void kinematics_update(double gyroX, double gyroY, double gyroZ, double accelX, double accelY, double accelZ) {
+void kinematics_update(float gyroX, float gyroY, float gyroZ, float accelX, float accelY, float accelZ) {
 
     // Normalize accel values
-    double norm = invSqrt(accelX * accelX + accelY * accelY + accelZ * accelZ);
+    float norm = invSqrt(accelX * accelX + accelY * accelY + accelZ * accelZ);
     accelX /= norm;
     accelY /= norm;
     accelZ /= norm;
@@ -24,21 +24,21 @@ void kinematics_update(double gyroX, double gyroY, double gyroZ, double accelX, 
     bool orientation = true; // up-side UP
     if (accelZ < 0.00) orientation = false; // up-side DOWN    
     
-    double accelXangle = atan2(accelY, accelZ);
-    double accelYangle = atan2(accelX, accelZ); 
-    //double accelYangle = atan2(*accelX, sqrt(*accelY * *accelY + *accelZ * *accelZ));   
+    float accelXangle = atan2(accelY, accelZ);
+    float accelYangle = atan2(accelX, accelZ); 
+    //float accelYangle = atan2(*accelX, sqrt(*accelY * *accelY + *accelZ * *accelZ));   
     
     // Accelerometer cut-off
-    double accelWeight = 0.0025; // normal operation
+    float accelWeight = 0.0025; // normal operation
     if (norm > 13.0 || norm < 7.0) accelWeight = 0.00; // gyro only
     
     // Save current time into variable for better computation time
     unsigned long now = micros();    
     
     // Fuse in gyroscope
-    kinematicsAngle[XAXIS] = kinematicsAngle[XAXIS] + (gyroX * (double)(now - kinematics_timer) / 1000000);
-    kinematicsAngle[YAXIS] = kinematicsAngle[YAXIS] + (gyroY * (double)(now - kinematics_timer) / 1000000);
-    kinematicsAngle[ZAXIS] = kinematicsAngle[ZAXIS] + (gyroZ * (double)(now - kinematics_timer) / 1000000);  
+    kinematicsAngle[XAXIS] = kinematicsAngle[XAXIS] + (gyroX * (float)(now - kinematics_timer) / 1000000);
+    kinematicsAngle[YAXIS] = kinematicsAngle[YAXIS] + (gyroY * (float)(now - kinematics_timer) / 1000000);
+    kinematicsAngle[ZAXIS] = kinematicsAngle[ZAXIS] + (gyroZ * (float)(now - kinematics_timer) / 1000000);  
     
     // Normalize gyro kinematics (+ - PI)
     if (kinematicsAngle[XAXIS] > PI) kinematicsAngle[XAXIS] -= TWO_PI;
